quad: preserve proper quad axes, impl random, pdf
The prior version modified the axis vectors to be the axis divided by
the square of its length, to make it easier to generate the Barycentric
coordinates of the ray-plane intersection through the dot product.

Unfortunately, this messes up other calculations that actually need the
unaltered axes, so this change stores the scaled axes as `measure_A` and
`measure_B`.

In The Rest of Your Life, implements the `pdf_value()` and `random()`
methods.